fix: migrator related tests fix: deploy storybook if dashboard gets deployed
