+++ .gitignore
@@ -65,7 +65,8 @@ handshake-nodejs-reporters/packages/handshake-dashboard/public/antd.min.css
 /handshake-nodejs-reporters/packages/handshake-dashboard/test-export.ps1
 
 handshake-nodejs-reporters/packages/handshake-dashboard/analyze/*
-
+TestMigrationRollback
+TestReversionRollback
 BackUpTestReports
 TestMigration
 handshake/dashboard/
