+++ __test__/test_patch_jobs/test_migration/test_scripts.py
@@ -135,56 +135,3 @@ class TestMigrationScripts:
         await assert_migration(7, 6, MigrationStatus.PASSED, MigrationTrigger.CLI)
         version = await ConfigBase.filter(key=ConfigKeys.version).first()
         assert int(version.value) == 6
-
-
-async def test_rollback_revert(get_vth_connection, db_path, root_dir):
-    # we would be causing an error while executing the reversion
-    # testing to see if the changes are committed or not (expectation: rollback)
-    connection = await get_vth_connection(db_path, OLDEST_VERSION)
-    await connection.execute_query(
-        "drop table taskbase;",
-    )
-    # we will run the wrong reversion script and see what would happen 😈
-
-    # we are in the oldest version: v5, but we are trying to revert v6
-    assert not revert_step_back(
-        OLDEST_VERSION + 1,
-        db_path,
-    ), "it should have rolled back"
-
-    version = await ConfigBase.filter(key=ConfigKeys.version).first()
-    assert (
-        int(version.value) == OLDEST_VERSION
-    )  # no changes were made apart from corruption I did
-
-    record = await assert_migration(
-        OLDEST_VERSION + 1,
-        OLDEST_VERSION,
-        MigrationStatus.FAILED,
-        MigrationTrigger.CLI,
-    )
-    assert "no such table: taskbase" in record.error
-
-
-async def test_rollback_migration(get_vth_connection, db_path, root_dir):
-    # we would be causing an error while executing the migration
-    # testing to see if the changes are committed or not (expectation: rollback)
-    connection = await get_vth_connection(db_path, OLDEST_VERSION)
-    await connection.execute_query(
-        "drop table taskbase;",
-    )
-    # we corrupt the database and see what would happen 😈
-
-    assert not migration(
-        db_path,
-    ), "it should have rolled back"
-
-    version = await ConfigBase.filter(key=ConfigKeys.version).first()
-    assert (
-        int(version.value) == OLDEST_VERSION
-    )  # no changes were made apart from corruption I did
-
-    record = await assert_migration(
-        OLDEST_VERSION, DB_VERSION, MigrationStatus.FAILED, MigrationTrigger.AUTOMATIC
-    )
-    assert "no such table: taskbase" in record.error
