+++ handshake-nodejs-reporters/packages/common-handshakes/.version
@@ -1,1 +1,1 @@
-{"0":{"url":"https://api.github.com/repos/RahulARanger/handshake/releases/assets/169496079","id":169496079,"node_id":"RA_kwDOKgpJeM4KGk4P","name":"dashboard.tar.bz2","label":"","content_type":"application/x-bzip2","state":"uploaded","size":802769,"download_count":0,"created_at":"2024-05-22T21:17:08Z","updated_at":"2024-05-22T21:17:09Z","browser_download_url":"https://github.com/RahulARanger/handshake/releases/download/0.6.3/dashboard.tar.bz2"},"version":"0.3.8"}
+{"0": {"url": "https://api.github.com/repos/RahulARanger/handshake/releases/assets/169496079", "id": 169496079, "node_id": "RA_kwDOKgpJeM4KGk4P", "name": "dashboard.tar.bz2", "label": "", "content_type": "application/x-bzip2", "state": "uploaded", "size": 802769, "download_count": 0, "created_at": "2024-05-22T21:17:08Z", "updated_at": "2024-05-22T21:17:09Z", "browser_download_url": "https://github.com/RahulARanger/handshake/releases/download/0.6.3/dashboard.tar.bz2"}, "version": "0.4.0"}
