+++ handshake/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.3.8"
+__version__ = "0.4.0"
